+++ packages/contracts/deploy/40_finalize-managing-dao/20_register-managing-dao-on-dao-registry.ts
@@ -1,0 +1,40 @@
+import {HardhatRuntimeEnvironment} from 'hardhat/types';
+import {DeployFunction} from 'hardhat-deploy/types';
+
+import {getContractAddress} from '../helpers';
+
+const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
+  const {getNamedAccounts, ethers} = hre;
+  const {deployer} = await getNamedAccounts();
+
+  // Get info from .env
+  const daoSubdomain = process.env.MANAGINGDAO_SUBDOMAIN || '';
+
+  if (!daoSubdomain)
+    throw new Error('ManagingDAO subdomain has not been set in .env');
+
+  // Get `managingDAO` address.
+  const managingDAOAddress = await getContractAddress('DAO', hre);
+
+  // Get `DAORegistry` address.
+  const daoRegistryAddress = await getContractAddress('DAORegistry', hre);
+
+  // Get `DAORegistry` contract.
+  const daoRegistryContract = await ethers.getContractAt(
+    'DAORegistry',
+    daoRegistryAddress
+  );
+
+  // Register `managingDAO` on `DAORegistry`.
+  const registerTx = await daoRegistryContract.register(
+    managingDAOAddress,
+    deployer,
+    daoSubdomain
+  );
+  await registerTx.wait();
+  console.log(
+    `Registered the (managingDAO: ${managingDAOAddress}) on (DAORegistry: ${daoRegistryAddress}), see (tx: ${registerTx.hash})`
+  );
+};
+export default func;
+func.tags = ['RegisterManagingDAO'];