+++ packages/contracts/src/core/plugin/Plugin.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: MIT
+// SPDX-License-Identifier: AGPL-3.0-or-later
 
 pragma solidity 0.8.17;
 
@@ -10,7 +10,7 @@ import {IPlugin} from "./IPlugin.sol";
 
 /// @title Plugin
 /// @author Aragon Association - 2022-2023
-/// @notice An abstract, non-upgradeable inherit from when creating a plugin being deployed via the `new` keyword.
+/// @notice An abstract, non-upgradeable contract to inherit from when creating a plugin being deployed via the `new` keyword.
 abstract contract Plugin is IPlugin, ERC165, DaoAuthorizable {
     /// @notice Constructs the plugin by storing the associated DAO.
     /// @param _dao The DAO contract.