+++ packages/contracts/src/core/plugin/PluginUUPSUpgradeable.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: MIT
+// SPDX-License-Identifier: AGPL-3.0-or-later
 
 pragma solidity 0.8.17;
 
@@ -20,6 +20,11 @@ abstract contract PluginUUPSUpgradeable is
     DaoAuthorizableUpgradeable
 {
     // NOTE: When adding new state variables to the contract, the size of `_gap` has to be adapted below as well.
+
+    /// @notice Disables the initializers on the implementation contract to prevent it from being left uninitialized.
+    constructor() {
+        _disableInitializers();
+    }
 
     /// @inheritdoc IPlugin
     function pluginType() public pure override returns (PluginType) {
@@ -46,9 +51,9 @@ abstract contract PluginUUPSUpgradeable is
     }
 
     /// @notice Returns the address of the implementation contract in the [proxy storage slot](https://eips.ethereum.org/EIPS/eip-1967) slot the [UUPS proxy](https://eips.ethereum.org/EIPS/eip-1822) is pointing to.
-    /// @return implementation The address of the implementation contract.
-    function getImplementationAddress() public view returns (address implementation) {
-        implementation = _getImplementation();
+    /// @return The address of the implementation contract.
+    function implementation() public view returns (address) {
+        return _getImplementation();
     }
 
     /// @notice Internal method authorizing the upgrade of the contract via the [upgradeabilty mechanism for UUPS proxies](https://docs.openzeppelin.com/contracts/4.x/api/proxy#UUPSUpgradeable) (see [ERC-1822](https://eips.ethereum.org/EIPS/eip-1822)).