+++ packages/contracts/src/framework/plugin/repo/PluginRepo.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier:    MIT
+// SPDX-License-Identifier:    AGPL-3.0-or-later
 
 pragma solidity 0.8.17;
 
@@ -121,6 +121,7 @@ contract PluginRepo is
         __PermissionManager_init(initialOwner);
 
         _grant(address(this), initialOwner, MAINTAINER_PERMISSION_ID);
+        _grant(address(this), initialOwner, UPGRADE_REPO_PERMISSION_ID);
     }
 
     /// @inheritdoc IPluginRepo