+++ packages/contracts/src/framework/plugin/setup/PluginSetupProcessor.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: MIT
+// SPDX-License-Identifier: AGPL-3.0-or-later
 
 pragma solidity 0.8.17;
 
@@ -521,9 +521,8 @@ contract PluginSetupProcessor {
             _params.pluginSetupRef.versionTag
         );
 
-        address currentImplementation = PluginUUPSUpgradeable(_params.plugin)
-            .getImplementationAddress();
-        address newImplementation = PluginSetup(version.pluginSetup).getImplementationAddress();
+        address currentImplementation = PluginUUPSUpgradeable(_params.plugin).implementation();
+        address newImplementation = PluginSetup(version.pluginSetup).implementation();
 
         if (currentImplementation != newImplementation) {
             _upgradeProxy(_params.plugin, newImplementation, _params.initData);