+++ packages/contracts/src/framework/utils/ens/ENSMigration.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: MIT
+// SPDX-License-Identifier: AGPL-3.0-or-later
 // This is a migration file as suggested here https://docs.ens.domains/deploying-ens-on-a-private-chain#migration-file-example
 // to compile the contracts and make their artifacts available in our tests.