+++ packages/contracts/src/plugins/governance/majority-voting/token/TokenVoting.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: MIT
+// SPDX-License-Identifier: AGPL-3.0-or-later
 
 pragma solidity 0.8.17;
 
@@ -27,11 +27,6 @@ contract TokenVoting is IMembership, MajorityVotingBase {
 
     /// @notice Thrown if the voting power is zero
     error NoVotingPower();
-
-    /// @dev Used to disallow initializing the implementation contract by an attacker for extra safety.
-    constructor() {
-        _disableInitializers();
-    }
 
     /// @notice Initializes the component.
     /// @dev This method is required to support [ERC-1822](https://eips.ethereum.org/EIPS/eip-1822).