+++ packages/subgraph/tests/multisig/utils.ts
@@ -271,6 +271,7 @@ export function createMultisigProposalEntityState(
   minApprovals: string = TWO,
   startDate: string = START_DATE,
   endDate: string = END_DATE,
+  executable: boolean = false,
   executed: boolean = false,
   allowFailureMap: string = ALLOW_FAILURE_MAP,
 
@@ -286,9 +287,10 @@ export function createMultisigProposalEntityState(
   multisigProposal.creator = Address.fromString(creator);
   multisigProposal.startDate = BigInt.fromString(startDate);
   multisigProposal.endDate = BigInt.fromString(endDate);
+  multisigProposal.executable = executable;
   multisigProposal.executed = executed;
   multisigProposal.snapshotBlock = BigInt.fromString(snapshotBlock);
-  multisigProposal.minApprovals = BigInt.fromString(minApprovals);
+  multisigProposal.minApprovals = BigInt.fromString(minApprovals).toI32();
   multisigProposal.allowFailureMap = BigInt.fromString(allowFailureMap);
   multisigProposal.createdAt = BigInt.fromString(createdAt);
   multisigProposal.creationBlockNumber = creationBlockNumber;